move setAdTestcookie to own module. Call in both DCR and Frontend commercial bundles
