+++ static/src/javascripts/bootstraps/commercial.dcr.js
@@ -5,6 +5,7 @@ import config from 'lib/config';
 import { catchErrorsWithContext } from 'lib/robust';
 import { markTime } from 'lib/user-timing';
 import reportError from 'lib/report-error';
+import { init as setAdTestCookie } from 'commercial/modules/set-adtest-cookie';
 import { init as initHighMerch } from 'commercial/modules/high-merch';
 import { init as initArticleAsideAdverts } from 'commercial/modules/article-aside-adverts';
 import { init as initArticleBodyAdverts } from 'commercial/modules/article-body-adverts';
@@ -30,6 +31,7 @@ import { init as initRedplanet } from 'commercial/modules/dfp/redplanet';
 import {refresh as refreshUserFeatures} from "common/modules/commercial/user-features";
 
 const commercialModules: Array<Array<any>> = [
+    ['cm-setAdTestCookie', setAdTestCookie],
     ['cm-adFreeSlotRemove', adFreeSlotRemove],
     ['cm-closeDisabledSlots', closeDisabledSlots],
     ['cm-comscore', initComscore],
